Extract Django settings script to standalone .py file
Move the inline Python script from django.ts into
vc_django_settings.py, matching the pattern used by vc_init_dev.py.
Read it at module load time and pass via python -c to keep sys.path[0]
as the project directory.
